requirements.txt
accelerate==1.4.0
acres==0.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
Authlib==1.5.2
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
dataclasses-json==0.6.7
deepsearch-glm==1.0.0
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
docling==2.15.1
docling-core==2.15.1
docling-ibm-models==3.2.1
docling-parse==3.1.1
durationpy==0.9
easyocr==1.7.2
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.1.21
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.14.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.5
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kor==3.0.0
kubernetes==32.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-docling==0.2.0
langchain-google-genai==2.0.9
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langchain-unstructured==0.1.6
langsmith==0.3.0
lark==1.2.2
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nipype==1.9.2
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.19.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
posthog==3.13.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
prov==2.0.1
psutil==7.0.0
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.1
pymongo==4.12.0
PyMuPDF==1.25.2
pyparsing==3.2.1
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
pywin32==307
pyxnat==1.6.2
PyYAML==6.0.2
ratelimit==2.2.1
rdflib==6.3.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
ruff==0.11.3
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
semchunk==2.2.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tesseract==0.1.3
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traits==6.4.3
transformers==4.49.0
typer==0.12.5
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured-client==0.29.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
